requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-JWT==0.2.0
Flask-Mail==0.9.0
Flask-SQLAlchemy==1.0
Flask-WTF==0.10.0
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.7
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.6.5
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
blinker==1.3
bouncer==0.1.10
celery==3.1.13
click==2.5
cssmin==0.2.0
factory-boy==2.4.1
flask-bouncer==0.1.13
gunicorn==19.1.0
itsdangerous==0.24
jsmin==2.0.11
kombu==3.0.21
mock==1.0.1
nose==1.3.3
psycopg2==2.5.3
py-bcrypt==0.4
pytz==2014.4
redis==2.10.1
requests==2.3.0
webassets==0.10.1
